refactor: extract grid row computation to separate utility function
