+++ apps/web/app/components/blocks/structure/MultiGrid/types.ts
@@ -1,5 +1,23 @@
 import type { Block } from '@plentymarkets/shop-api';
 
+export interface GridCell {
+  colIndex: number;
+  span: number;
+  hasContent: boolean;
+}
+
+export interface GridRow {
+  cells: GridCell[];
+  free: number;
+}
+
+export interface DragState {
+  colIndex: number;
+  startX: number;
+  startSpan: number;
+  unitW: number;
+}
+
 interface MultiGridLayout {
   marginTop?: number;
   marginBottom?: number;
